Move AssetValueSnapshot and DividendValue to ValueObjects
Both are immutable records with no identity (market data captured from
an external source), so they belong in ValueObjects rather than
Entities, which is reserved for aggregates with identity and lifecycle
(Asset, Broker, Investments).
