+++ Financial.Infrastructure/Services/AssetPriceService.cs
@@ -1,6 +1,7 @@
 using Financial.Application.DTOs;
 using Financial.Application.Interfaces;
 using Financial.Domain.Entities;
+using Financial.Domain.ValueObjects;
 using Financial.Infrastructure.Integrations.WebPageParser;
 
 namespace Financial.Infrastructure.Services;
