+++ Financial.Infrastructure/Services/DividendDataSourceAdapter.cs
@@ -1,5 +1,5 @@
 using Financial.Application.Interfaces;
-using Financial.Domain.Entities;
+using Financial.Domain.ValueObjects;
 using Financial.Infrastructure.Integrations.WebPageParser;
 
 namespace Financial.Infrastructure.Services;
